+++ lib/src/commands/get_tags.dart
@@ -27,6 +27,7 @@ class GetTags extends GgGitBase<List<String>> {
   Future<List<String>> exec({
     required Directory directory,
     required GgLog ggLog,
+    Map<String, dynamic> options = const {},
   }) async {
     final headOnly = argResults!['head-only'] as bool;
     late List<String> result;