+++ lib/src/commands/is_committed.dart
@@ -28,6 +28,7 @@ class IsCommitted extends GgGitBase<bool> {
   Future<bool> exec({
     required Directory directory,
     required GgLog ggLog,
+    Map<String, dynamic> options = const {},
   }) async {
     final messages = <String>[];