+++ lib/src/commands/is_eol_lf.dart
@@ -29,6 +29,7 @@ class IsEolLf extends GgGitBase<bool> {
   Future<bool> exec({
     required Directory directory,
     required GgLog ggLog,
+    Map<String, dynamic> options = const {},
   }) async {
     final messages = <String>[];